March4_finalAssignment
Turning in final assignment for Lab 1, Geography 575.

data exploration map with the total population for the 31 largest metro areas in the U.S. from 2005-2017. Intended to do % of population aged 25-34, but the data didn't change enough to register in proportional symbols and had to settle for something else in the dataset that I had handy, which was only the total population.